bug: add disable to prev and next button
