+++ lib/constants/carouselData.tsx
@@ -1,0 +1,41 @@
+import MentorImageOne from "@/public/assets/images/landing-page/mentorTypes-1.png";
+import MentorImageTwo from "@/public/assets/images/landing-page/mentorTypes-2.png";
+import MentorImageThree from "@/public/assets/images/landing-page/mentorTypes-3.png";
+import MentorImageFour from "@/public/assets/images/landing-page/mentorTypes-4.png";
+
+const imgCarousel = [
+  {
+    id: 1,
+    img: MentorImageTwo,
+  },
+  {
+    id: 2,
+    img: MentorImageThree,
+  },
+  {
+    id: 3,
+    img: MentorImageFour,
+  },
+  {
+    id: 4,
+    img: MentorImageThree,
+  },
+  {
+    id: 5,
+    img: MentorImageOne,
+  },
+  {
+    id: 6,
+    img: MentorImageTwo,
+  },
+  {
+    id: 7,
+    img: MentorImageOne,
+  },
+  {
+    id: 8,
+    img: MentorImageTwo,
+  },
+];
+
+export default imgCarousel;
